:recycle: move repository to infra
