+++ integration_tests/repository/contest_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/traPtitech/traPortfolio/integration_tests/testutils"
 	"github.com/traPtitech/traPortfolio/interfaces/external/mock_external_e2e"
 
-	irepository "github.com/traPtitech/traPortfolio/interfaces/repository"
+	irepository "github.com/traPtitech/traPortfolio/infrastructure/repository"
 	urepository "github.com/traPtitech/traPortfolio/usecases/repository"
 	"github.com/traPtitech/traPortfolio/util/mockdata"
 	"github.com/traPtitech/traPortfolio/util/random"
