+++ util/mockdata/model.go
@@ -4,8 +4,8 @@ import (
 	"time"
 
 	"github.com/traPtitech/traPortfolio/domain"
+	"github.com/traPtitech/traPortfolio/infrastructure/repository/model"
 	"github.com/traPtitech/traPortfolio/interfaces/database"
-	"github.com/traPtitech/traPortfolio/interfaces/repository/model"
 )
 
 var (
